backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-dotenv
python-multipart
openai
psycopg2-binary
SQLAlchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
typing_extensions
pydantic[email]